More options and fix type misdeserialization bug
Bug: Flink infers types, and using 0 instead of 0.0 causes Flink to
think all of the values in that field are longs, causing a floating
point value to be mis-deserialized.

Options:

* On request, use an implementation of the java RNG in python. At
the moment, because the Python API runs the script selectively for each
operation, the RNG resets each time. Using either the Python native RNG
or the Java RNG with a seed will result in the same random u vector
being generated with each atom.

* Verbose (don't display useful information at initialization). This
cleans up the stdout log if we need to output something for debugging
purposes.

Termination criterion is not working at the moment; results seem valid
but require max iterations to finish.